Drop audiogen.
audiogen wasn't reliable - sometimes it would work and sometimes
it wouldn't, resulting in audio playing too quickly and lots of
underruns.

This new approach seems more reliable, although it also has problems
at the start of play if bitrate is set too high, and I'm not sure
the sine wave generation code is spot on yet.